[#113] move database.go under simra package
