+++ simra/database.go
@@ -1,4 +1,4 @@
-package database
+package simra
 
 import "fmt"
